web_demo/requirements.txt
# ==============================================================================
# Production Dependencies for KrishiSahayak Web Demo
# ==============================================================================
# This file contains only the essential packages needed to run the application.
# Install with: pip install -r requirements.txt
# ==============================================================================

# ------------------------------------------------------------------------------
# Core ML/AI Framework
# ------------------------------------------------------------------------------
torch==2.3.1
transformers==4.52.4
bitsandbytes==0.43.1
accelerate==0.31.0
safetensors==0.5.3

# ------------------------------------------------------------------------------
# Vector Database & Search
# ------------------------------------------------------------------------------
faiss-cpu==1.8.0
sentence-transformers==3.0.1

# ------------------------------------------------------------------------------
# Audio Processing
# ------------------------------------------------------------------------------
openai-whisper==20231117
gTTS==2.5.1
pydub==0.25.1

# ------------------------------------------------------------------------------
# Web Interface & Server
# ------------------------------------------------------------------------------
gradio==4.37.2
fastapi==0.115.14
uvicorn==0.35.0
python-multipart==0.0.20

# ------------------------------------------------------------------------------
# Data Processing & Utilities
# ------------------------------------------------------------------------------
pandas==2.2.2
numpy==1.26.4
scipy==1.14.0
Pillow==10.4.0
tqdm==4.67.1